refactor: rename getLocalStorageItems to localStorage and update imports for consistency
