+++ client/src/utils/getDefaultEndpoint.ts
@@ -4,16 +4,20 @@ import type {
   TEndpointsConfig,
   EModelEndpoint,
 } from 'librechat-data-provider';
-import getLocalStorageItems from './getLocalStorageItems';
+import { getLocalStorageItems } from './localStorage';
 import { mapEndpoints } from './endpoints';
 
 type TConvoSetup = Partial<TPreset> | Partial<TConversation>;
 
 type TDefaultEndpoint = { convoSetup: TConvoSetup; endpointsConfig: TEndpointsConfig };
 
-const getEndpointFromSetup = (convoSetup: TConvoSetup, endpointsConfig: TEndpointsConfig) => {
-  const { endpoint: targetEndpoint } = convoSetup || {};
-  if (targetEndpoint && endpointsConfig?.[targetEndpoint ?? '']) {
+const getEndpointFromSetup = (
+  convoSetup: TConvoSetup | null,
+  endpointsConfig: TEndpointsConfig,
+) => {
+  let { endpoint: targetEndpoint = '' } = convoSetup || {};
+  targetEndpoint = targetEndpoint ?? '';
+  if (targetEndpoint && endpointsConfig?.[targetEndpoint]) {
     return targetEndpoint;
   } else if (targetEndpoint) {
     console.warn(`Illegal target endpoint ${targetEndpoint} ${endpointsConfig}`);
@@ -24,8 +28,8 @@ const getEndpointFromSetup = (convoSetup: TConvoSetup, endpointsConfig: TEndpoin
 const getEndpointFromLocalStorage = (endpointsConfig: TEndpointsConfig) => {
   try {
     const { lastConversationSetup } = getLocalStorageItems();
-    const { endpoint } = lastConversationSetup;
-    const isDefaultConfig = Object.values(endpointsConfig ?? {})?.every((value) => !value);
+    const { endpoint } = lastConversationSetup ?? { endpoint: null };
+    const isDefaultConfig = Object.values(endpointsConfig ?? {}).every((value) => !value);
 
     if (isDefaultConfig && endpoint) {
       return endpoint;
@@ -35,7 +39,7 @@ const getEndpointFromLocalStorage = (endpointsConfig: TEndpointsConfig) => {
       return endpoint;
     }
 
-    return endpoint && endpointsConfig?.[endpoint ?? ''] ? endpoint : null;
+    return endpoint && endpointsConfig?.[endpoint] != null ? endpoint : null;
   } catch (error) {
     console.error(error);
     return null;
@@ -47,9 +51,12 @@ const getDefinedEndpoint = (endpointsConfig: TEndpointsConfig) => {
   return endpoints.find((e) => Object.hasOwn(endpointsConfig ?? {}, e));
 };
 
-const getDefaultEndpoint = ({ convoSetup, endpointsConfig }: TDefaultEndpoint): EModelEndpoint => {
+const getDefaultEndpoint = ({
+  convoSetup,
+  endpointsConfig,
+}: TDefaultEndpoint): EModelEndpoint | string | undefined => {
   return (
-    getEndpointFromSetup(convoSetup, endpointsConfig) ||
+    (getEndpointFromSetup(convoSetup, endpointsConfig) ?? '') ||
     getEndpointFromLocalStorage(endpointsConfig) ||
     getDefinedEndpoint(endpointsConfig)
   );
