+++ client/src/utils/index.ts
@@ -12,6 +12,7 @@ export * from './messages';
 export * from './languages';
 export * from './endpoints';
 export * from './sharedLink';
+export * from './localStorage';
 export * from './promptGroups';
 export { default as cn } from './cn';
 export { default as logger } from './logger';
@@ -21,7 +22,6 @@ export { default as cleanupPreset } from './cleanupPreset';
 export { default as validateIframe } from './validateIframe';
 export { default as buildDefaultConvo } from './buildDefaultConvo';
 export { default as getDefaultEndpoint } from './getDefaultEndpoint';
-export { default as getLocalStorageItems } from './getLocalStorageItems';
 
 export const languages = [
   'java',
